Create password generator screen and components
GeneratorView is the main view that will contains:
PasswordGeneratedView, the result of all combintations;
PasswordSizeView, to set a password size;
CharInclusionView, to enable some char combinations;
SeparatorsView, to choose separators;
SaveView, the footer component with save flow.

For now, the views contains only local states, and doesn't cause side effects on the app.

Signed-off-by: Ronan Nunes <ronan@uber.com>